perf: download and cache polyfills on-the-fly
