+++ src/service-worker.js
@@ -4,6 +4,11 @@ import {
   routes as __routes__
 } from '../__sapper__/service-worker.js'
 
+import {
+  get,
+  post
+} from './routes/_utils/ajax'
+
 const timestamp = process.env.SAPPER_TIMESTAMP
 const ASSETS = `assets_${timestamp}`
 const WEBPACK_ASSETS = `webpack_assets_${timestamp}`
@@ -17,7 +22,8 @@ const assets = __assets__
 // `shell` is an array of all the files generated by webpack
 // also contains '/index.html' for some reason
 const webpackAssets = __shell__
-  .filter(filename => !filename.endsWith('.map'))
+  .filter(filename => !filename.endsWith('.map')) // don't bother with sourcemaps
+  .filter(filename => !filename.includes('$polyfill$')) // polyfills are cached on-demand
 
 // `routes` is an array of `{ pattern: RegExp }` objects that
 // match the pages in your src
@@ -92,6 +98,13 @@ self.addEventListener('fetch', event => {
         if (response) {
           return response
         }
+      }
+      // for polyfills, cache them on-the-fly
+      if (url.pathname.includes('$polyfill$')) {
+        let response = await fetch(req)
+        // cache asynchronously, don't wait
+        caches.open(WEBPACK_ASSETS).then(cache => cache.put(req, response))
+        return response.clone()
       }
     }
 
@@ -247,62 +260,3 @@ self.addEventListener('notificationclick', event => {
     }
   })())
 })
-
-// Copy-paste from ajax.js
-async function get (url, headers, options) {
-  return _fetch(url, makeFetchOptions('GET', headers), options)
-}
-
-async function post (url, body, headers, options) {
-  return _putOrPostOrPatch('POST', url, body, headers, options)
-}
-
-async function _putOrPostOrPatch (method, url, body, headers, options) {
-  let fetchOptions = makeFetchOptions(method, headers)
-  if (body) {
-    if (body instanceof FormData) {
-      fetchOptions.body = body
-    } else {
-      fetchOptions.body = JSON.stringify(body)
-      fetchOptions.headers['Content-Type'] = 'application/json'
-    }
-  }
-  return _fetch(url, fetchOptions, options)
-}
-
-async function _fetch (url, fetchOptions, options) {
-  let response
-  if (options && options.timeout) {
-    response = await fetchWithTimeout(url, fetchOptions, options.timeout)
-  } else {
-    response = await fetch(url, fetchOptions)
-  }
-  return throwErrorIfInvalidResponse(response)
-}
-
-async function throwErrorIfInvalidResponse (response) {
-  let json = await response.json()
-  if (response.status >= 200 && response.status < 300) {
-    return json
-  }
-  if (json && json.error) {
-    throw new Error(response.status + ': ' + json.error)
-  }
-  throw new Error('Request failed: ' + response.status)
-}
-
-function fetchWithTimeout (url, fetchOptions, timeout) {
-  return new Promise((resolve, reject) => {
-    fetch(url, fetchOptions).then(resolve, reject)
-    setTimeout(() => reject(new Error(`Timed out after ${timeout / 1000} seconds`)), timeout)
-  })
-}
-
-function makeFetchOptions (method, headers) {
-  return {
-    method,
-    headers: Object.assign(headers || {}, {
-      'Accept': 'application/json'
-    })
-  }
-}
